# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API fast-learning",
        "description": "Bienvenue sur la version 1 de la documentation API de fast-learning.<br><br><u>Condition préalable :</u><br><li>Avoir un compte utilisateur</li><br><li>Une clé d'application <b>apiKey</b> vous sera donné lorsque que vous vous connectez. Utilisé là pour faire des requêtes</li><br>",
        "contact": {
            "name": ": TODO AFTER THE SERVER IS ONLINE",
            "email": "elian.yutz@hotmail.fr"
        },
        "license": {
            "name": "TODO AFTER THE SERVER IS ONLINE",
            "url": "https:google.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:80/api",
            "description": "Serveur API Rest"
        }
    ],
    "paths": {
...
        "/folders": {
            "get": {
                "tags": [
                    "Dossier"
                ],
                "summary": "Permet de récupérer l'ensemble des dossiers",
                "description": "Permet de récupérer l'ensemble des dossiers",
                "operationId": "folders",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Dossier"
                ],
                "summary": "création d'un dossier",
                "description": "Store a newly created folder in storage.",
                "operationId": "initFolder",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "",
                                        "type": "string",
                                        "example": "Dossier1"
                                    },
                                    "content": {
                                        "description": "",
                                        "type": "string",
                                        "example": "Content of the folder"
                                    },
                                    "is_public": {
                                        "description": "",
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "parent_id": {
                                        "description": "",
                                        "type": "integer",
                                        "example": null
                                    },
                                    "created_by_user": {
                                        "description": "",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/folders/{id}": {
            "get": {
                "tags": [
                    "Dossier"
                ],
                "summary": "Permet de récupérer un dossier",
                "description": "Permet de récupérer un dossier",
                "operationId": "folder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id du dossier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Dossier"
                ],
                "summary": "Permet de supprimer un dossier",
                "description": "Permet de supprimer un dossier",
                "operationId": "deleteFolder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id du dossier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Dossier"
                ],
                "summary": "Permet de mettre à jour un dossier",
                "description": "Permet de mettre à jour un dossier",
                "operationId": "updateFolder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id du dossier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "",
                                        "type": "string",
                                        "example": "Nouveau dossier 1"
                                    },
                                    "content": {
                                        "description": "",
                                        "type": "string",
                                        "example": "Content of the folder"
                                    },
                                    "is_public": {
                                        "description": "",
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "parent_id": {
                                        "description": "",
                                        "type": "integer",
                                        "example": null
                                    },
                                    "type": {
                                        "description": "type",
                                        "type": "string",
                                        "example": "type"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}